Text IntoElement, imports, updated counter example
