+++ crates/zoon/src/element.rs
@@ -63,111 +63,6 @@ impl<'a, T: Element> IntoElement<'a> for T {
     }
 }
 
-impl<'a> IntoElement<'a> for String {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self)
-    }
-}
-
-impl<'a> IntoElement<'a> for &'a str {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self)
-    }
-}
-
-impl<'a> IntoElement<'a> for Cow<'a, str> {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self)
-    }
-}
-
-impl<'a> IntoElement<'a> for u8 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for u16 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for u32 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for u64 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for u128 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for usize {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for i8 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for i16 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for i32 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for i64 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for i128 {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
-impl<'a> IntoElement<'a> for isize {
-    type EL = Text;
-    fn into_element(self) -> Self::EL {
-        Text::with_text(self.to_string())
-    }
-}
-
